issue 4 fix and squash
