+++ application.properties
@@ -1,4 +1,4 @@
 #Grails Metadata file
-#Fri Feb 08 09:47:45 CST 2013
-app.grails.version=2.1.0
+#Mon Mar 04 10:52:53 CST 2013
+app.grails.version=2.1.3
 app.name=coffeescript-compiler
